adjusted modal positioning
