+++ src/actions/bureauException.js
@@ -13,7 +13,7 @@ const dummyData = [
   {
     id: 1,
     Name: 'John Doe',
-    BureauNames: ['OBO', 'ENR'],
+    BureauNames: ['OBO', 'ENR', 'AO'],
     NameAbbreviation: 'JD',
   },
   {
