+++ src/sass/_bureauExceptionList.scss
@@ -24,7 +24,7 @@
     }
   }
   .checkbox-pac {
-      min-width: 450px;
+      min-width: 470px;
       text-align: left;
     }
     .checkbox-pos {
